+++ src/Commands/SystemCommands/NewchattitleCommand.php
@@ -14,6 +14,8 @@ use Longman\TelegramBot\Commands\SystemCommand;
 
 /**
  * New chat title command
+ *
+ * @todo Remove due to deprecation!
  */
 class NewchattitleCommand extends SystemCommand
 {
@@ -42,6 +44,8 @@ class NewchattitleCommand extends SystemCommand
     {
         //$message = $this->getMessage();
         //$new_chat_title = $message->getNewChatTitle();
+
+        trigger_error(__CLASS__ . ' is deprecated and will be removed and handled by ' . GenericmessageCommand::class . ' by default in a future release.', E_USER_DEPRECATED);
 
         return parent::execute();
     }