+++ CHANGELOG.md
@@ -6,8 +6,10 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 ## [Unreleased]
 ### Added
 ### Changed
+- All Message field types dynamically search for an existing Command class that can handle them.
 ### Deprecated
 - Botan.io service has been discontinued.
+- Most built-in System Commands will be handled by GenericmessageCommand by default in a future release and will require a custom implementation.
 ### Removed
 ### Fixed
 - Constraint errors in `/cleanup` command.